Add weighted primitive size calculations
Previously, the size of a PrimApp in the SSA IL was based solely on the
number of arguments; this change introduces a new `weight` associated
with each Prim which multiplies this value by an integral factor. A
command-line option, "-weight-prims", controls whether this weight value
is considered, and is currently set to false by default.

Right now the only primitives with meaningful weights are the
Word_*CheckP primitives which are tentatively assigned a weight of ~1 in
an attempt to counteract performance issues related to inlining. Future
tweaking of other primitive weights may also prove beneficial.